PDF and Rmd which builds it
For use by workshop attendees. Using the Rmd is not recommended -- I mean, you can try to build the pdf from the Rmd but if you're at an intro to R workshop you will probably fail to do so at this stage in your R development. Just use the pdf.